Use meaningful exit codes (#2)
The sender computes a SHA-256 hash of the data it sends, and sends
this hash in the EOF packets.

The receiver computes a SHA-256 hash of the data it receives. It
compares this hash to the hash included in the EOF packets. If the
hashes match, exit with zero. Otherwise, exit with a non-zero code.

Closes #1