Add payload encryption option to GUI (#32)
Users can now choose to use gpg to encrypt/decrypt data sent using
the GUI. When configured, this protects users if they accidentally
send data to the wrong computer.

There is now a "PGP Key ID" field in the settings tab. When the key
ID is present, the sender will use gpg to encrypt data before
pydiode sends it, and the receiver will use gpg to decrypt the data.

Closes #31