Improve Reliability of Two Computer Transfers (#44)
Most significantly, we refactored to replace asyncio with
threads, which we found significantly reduced packet
loss, especially on Ubuntu 24.04.

When testing with asyncio, we found that small packets
and STDIN throttling increased packet loss, and caused
packets to arrive out of order. Thus, we use padding to
ensure all packets have equal size. Also, we retransmit
packets to fill chunks, and when we are waiting for more
data.  Leaving these changes after the switch to threads
because they seem likely to help.

Fixes #41